#208 Добавлен CreatePostViewHolder

+++ app/src/main/java/io/golos/cyber_android/utils/Ints.kt
@@ -1,5 +1,5 @@
 package io.golos.cyber_android.utils
 
-fun Int?.positiveValue(): Long{
+fun Int?.positiveValue(): Int{
     return if(this != null && this >= 0) this else 0
 }
